+++ core/src/commonMain/kotlin/eu/kanade/tachiyomi/network/NetworkPreferences.kt
@@ -2,7 +2,12 @@ package eu.kanade.tachiyomi.network
 
 import eu.kanade.tachiyomi.core.preference.PreferenceStore
 
-class NetworkPreferences(private val preferenceStore: PreferenceStore) {
+class NetworkPreferences(
+    private val preferenceStore: PreferenceStore,
+    private val verboseLogging: Boolean,
+) {
+
+    fun verboseLogging() = preferenceStore.getBoolean("verbose_logging", verboseLogging)
 
     fun dohProvider() = preferenceStore.getInt("doh_provider", -1)